# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5eb64906bbf7210dd3bd2f89
  title: WeddingRing
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/user/{id}":
    delete:
      x-skaffolder-id: 5eb64907bbf7210dd3bd2f99
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5eb64906bbf7210dd3bd2f93
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles:
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5eb64907bbf7210dd3bd2f9b
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5eb64906bbf7210dd3bd2f93
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles:
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5eb64907bbf7210dd3bd2f97
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5eb64906bbf7210dd3bd2f93
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles:
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
...
  "/userprofiles/{id}":
    delete:
      x-skaffolder-id: 5eb64c47bbf7210dd3bd301e
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5eb64c33bbf7210dd3bd3015
      x-skaffolder-resource: UserProfiles
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles:
      summary: CRUD ACTION delete
      tags:
        - UserProfiles
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5eb64c46bbf7210dd3bd3019
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5eb64c33bbf7210dd3bd3015
      x-skaffolder-resource: UserProfiles
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles:
      summary: CRUD ACTION get
      tags:
        - UserProfiles
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfiles"
    post:
      x-skaffolder-id: 5eb64c46bbf7210dd3bd301a
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5eb64c33bbf7210dd3bd3015
      x-skaffolder-resource: UserProfiles
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles:
      summary: CRUD ACTION update
      tags:
        - UserProfiles
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfiles"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfiles"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      x-skaffolder-id: 5eb64906bbf7210dd3bd2f93
      x-skaffolder-id-db: 5eb64906bbf7210dd3bd2f8a
      x-skaffolder-id-entity: 5eb64906bbf7210dd3bd2f8b
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5eb64906bbf7210dd3bd2f8e
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5eb64906bbf7210dd3bd2f8f
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5eb64906bbf7210dd3bd2f8d
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5eb64906bbf7210dd3bd2f91
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5eb64906bbf7210dd3bd2f90
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5eb64906bbf7210dd3bd2f8c
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
    UserProfiles:
      x-skaffolder-id: 5eb64c33bbf7210dd3bd3015
      x-skaffolder-id-db: 5eb64906bbf7210dd3bd2f8a
      x-skaffolder-id-entity: 5eb64c32bbf7210dd3bd300e
      x-skaffolder-url: /userprofiles
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        AgeofEldestChild:
          type: number
          x-skaffolder-id-attr: 5ec0329fa940665206b0117c
          x-skaffolder-type: Number
        BirthPlace:
          type: string
          x-skaffolder-id-attr: 5ec0302ca940665206b01153
          x-skaffolder-type: String
        BirthYear:
          type: number
          x-skaffolder-id-attr: 5ec0302ca940665206b01154
          x-skaffolder-type: Number
        Budget:
          type: number
          x-skaffolder-id-attr: 5ec0329fa940665206b0117d
          x-skaffolder-type: Number
        Company:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b0116b
          x-skaffolder-type: String
        Complexion:
          type: string
          x-skaffolder-id-attr: 5ec0302ca940665206b01157
          x-skaffolder-type: String
        ContactNumber:
          type: number
          x-skaffolder-id-attr: 5ec0329fa940665206b01178
          x-skaffolder-type: Number
        CurrentCity:
          type: string
          x-skaffolder-id-attr: 5eb64c32bbf7210dd3bd3014
          x-skaffolder-type: String
          x-skaffolder-required: true
        CurrentState:
          type: string
          x-skaffolder-id-attr: 5ebff20ba940665206b00ff4
          x-skaffolder-type: String
        DOB:
          type: integer
          x-skaffolder-id-attr: 5eb64c32bbf7210dd3bd3012
          x-skaffolder-type: Date
          x-skaffolder-required: true
        Degree:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b01168
          x-skaffolder-type: String
        DegreePlace:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b01169
          x-skaffolder-type: String
        Education:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b01167
          x-skaffolder-type: String
        FIrstName:
          type: string
          x-skaffolder-id-attr: 5eb64c32bbf7210dd3bd300f
          x-skaffolder-type: String
          x-skaffolder-required: true
        Familytype:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b0117a
          x-skaffolder-type: String
        Father:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b0116e
          x-skaffolder-type: String
        FatherOccupation:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b0116f
          x-skaffolder-type: String
        Gotra:
          type: string
          x-skaffolder-id-attr: 5eb64c32bbf7210dd3bd3011
          x-skaffolder-type: String
        Grandfather:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b0116d
          x-skaffolder-type: String
        Height:
          type: number
          x-skaffolder-id-attr: 5ec0302ca940665206b01155
          x-skaffolder-type: Decimal
        LastName:
          type: string
          x-skaffolder-id-attr: 5eb64c32bbf7210dd3bd3010
          x-skaffolder-type: String
        Manglikstatus:
          type: boolean
          x-skaffolder-id-attr: 5ec0329fa940665206b01165
          x-skaffolder-type: Boolean
        Maritalstatus:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b01164
          x-skaffolder-type: String
          x-skaffolder-required: true
        MotherName:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b01170
          x-skaffolder-type: String
        MotherTounge:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b01176
          x-skaffolder-type: String
        MothersOccupation:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b01171
          x-skaffolder-type: String
        NumberOfChilderen:
          type: number
          x-skaffolder-id-attr: 5ec0329fa940665206b0117b
          x-skaffolder-type: Number
        NumberofBrothers:
          type: number
          x-skaffolder-id-attr: 5ec0329fa940665206b01172
          x-skaffolder-type: Number
        NumberofMarriedBrothers:
          type: number
          x-skaffolder-id-attr: 5ec0329fa940665206b01173
          x-skaffolder-type: Number
        NumberofMarriedBrothers:
          type: number
          x-skaffolder-id-attr: 5ec0329fa940665206b01175
          x-skaffolder-type: Number
        NumberofSisters:
          type: number
          x-skaffolder-id-attr: 5ec0329fa940665206b01174
          x-skaffolder-type: Number
        Occupation:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b0116a
          x-skaffolder-type: String
        Package:
          type: number
          x-skaffolder-id-attr: 5ec0329fa940665206b0116c
          x-skaffolder-type: Number
        PermanentAddress:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b01177
          x-skaffolder-type: String
        Physicalstatus:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b01166
          x-skaffolder-type: String
        Preferences:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b0117e
          x-skaffolder-type: String
        RelationWithContactPerson:
          type: string
          x-skaffolder-id-attr: 5ec0329fa940665206b01179
          x-skaffolder-type: String
        TimeOfBirth:
          type: number
          x-skaffolder-id-attr: 5eb64c32bbf7210dd3bd3013
          x-skaffolder-type: Decimal
        Weight:
          type: number
          x-skaffolder-id-attr: 5ec0302ca940665206b01156
          x-skaffolder-type: Number
      x-skaffolder-relations: 
      required:
        - _id
        - CurrentCity
        - DOB
        - FIrstName
        - Maritalstatus
...